Revert "made DELETE operation idempotent"
This reverts commit 535d8d25f98b723d2d577a7884abbdc8d04cf229.

+++ src/main/java/com/intern/userservice/service/impl/CardInfoServiceImpl.java
@@ -60,6 +60,9 @@ public class CardInfoServiceImpl implements CardInfoService {
             @CacheEvict(value = "userCards", allEntries = true)
     })
     public void deleteCardById(Long id) {
+        if (!cardInfoRepository.existsById(id)) {
+            throw new EntityNotFoundException("Card not found with id " + id);
+        }
         cardInfoRepository.deleteByIdNative(id);
     }
